okex: convert to TS
