+++ src/Market.ts
@@ -2,4 +2,5 @@ export type Market = {
     id: string;
     base: string;
     quote: string;
+    type?: string;
 };
